blockbuster/requirements.txt
appdirs==1.4.2
dj-database-url==0.4.2
Django==1.8
django-filter==1.0.1
django-webpack-loader==0.4.1
djangorestframework==3.5.4
factory-boy==2.8.1
Faker==0.7.9
gunicorn==19.6.0
ipaddress==1.0.18
Markdown==2.6.8
packaging==16.8
pyparsing==2.1.10
python-dateutil==2.6.0
six==1.10.0
whitenoise==3.3.0
